% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lordif_multiverse.R
\name{lordif_multiverse}
\alias{lordif_multiverse}
\title{Perform Multiverse Logistic Regression DIF Analysis}
\usage{
lordif_multiverse(items, group, verbose = FALSE)
}
\arguments{
\item{items}{A data frame or matrix of item responses, where rows represent
individuals and columns represent items.}

\item{group}{A vector indicating the group membership for each individual in
'items'. The length of 'group' must match the number of rows in 'items'.}

\item{verbose}{A logical value indicating whether to print detailed processing
messages. Defaults to FALSE.}
}
\value{
A data frame summarizing the DIF analysis results for each specification,
including the specification used (criterion and threshold) and the items identified
as exhibiting DIF.
}
\description{
This function conducts a multiverse analysis for differential item functioning (DIF)
using logistic regression. It automatically generates and tests multiple
specifications based on a range of criteria and thresholds to robustly assess DIF
across items.
}
\details{
The function tests for DIF using various criteria including Chi-squared,
Beta, Cox-Snell, Nagelkerke, and McFadden's R-squared, across a predefined range
of thresholds. It leverages the 'lordif' package for the underlying DIF analysis.
The 'dplyr', 'tidyr', and 'lapply' functions are used for data manipulation, and
'do.call' is used for dynamic function calls. Ensure that these packages are
installed and available in your R environment: 'lordif' for logistic regression DIF
analysis, 'dplyr' and 'tidyr' for data manipulation.
}
\examples{
# Use data from the lordif package
data(Anxiety, package = "lordif")
items_anxiety <- Anxiety[paste("R", 1:29, sep = "")]
group_age <- Anxiety$age

# Run multiverse analysis for different grouping variables
# The examples below are wrapped in \donttest{} to avoid long execution times
# during package checking. Remove \donttest{} if you want them to be executed
# during R CMD check.
\donttest{
multiverse_age <- lordif_multiverse(items_anxiety, group_age)
}

}
