% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lordif_multiverse.R
\name{plot_lordif_multiverse}
\alias{plot_lordif_multiverse}
\title{Plot Results of Multiverse Logistic Regression DIF Analysis}
\usage{
plot_lordif_multiverse(specification_results, items)
}
\arguments{
\item{specification_results}{A data frame containing the results of the DIF
analysis, including the specifications (criteria and thresholds) and the items
identified as exhibiting DIF.}

\item{items}{A data frame or matrix of item responses used in the DIF analysis.
This is used to set scale limits in the plots.}
}
\value{
The function invisibly returns a grid plot (cowplot object) combining a
histogram and a boxplot of the DIF analysis results. The plot is displayed but not
explicitly returned; use `return()` to capture the plot if needed.
}
\description{
This function generates visualizations for the results of a multiverse logistic
regression differential item functioning (DIF) analysis. It produces a histogram
showing the distribution of the number of items flagged for DIF across different
specifications and a boxplot summarizing the number of flagged items per
specification criterion and threshold.
}
\details{
The function first checks if all identified_items are missing. If so, it
assumes no items were flagged for DIF and adjusts the data accordingly. Otherwise,
it processes the `specification_results` to calculate the sum of flagged items for
each specification. It uses `ggplot2` for plotting and `cowplot` for arranging the
plots. Ensure `ggplot2`, `dplyr`, `tidyr`, and `cowplot` are installed and
available in your R environment.
}
\examples{
# Assuming 'specification_results' is the output from `lordif_multiverse` and
# 'items' is the original item response data:
data(Anxiety, package = "lordif")
items_anxiety <- Anxiety[paste("R", 1:29, sep = "")]

specification_results <- structure(list(
specification = c("Chisqr", "Chisqr", "Chisqr",
"Beta", "Beta", "CoxSnell", "CoxSnell", "CoxSnell", "CoxSnell",
"CoxSnell", "Nagelkerke", "Nagelkerke", "Nagelkerke", "Nagelkerke",
"Nagelkerke", "McFadden", "McFadden", "McFadden", "McFadden",
"McFadden"),
threshold = c(0.0001, 0.001, 0.01, 0.05, 0.1, 0.02,
              0.03, 0.04, 0.05, 0.06, 0.02, 0.03, 0.04, 0.05, 0.06, 0.02, 0.03,
              0.04, 0.05, 0.06),
identified_items = c(NA, "R9-R11-R24", "R9-R11-R18-R24",
              "R9-R11", NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,
              NA, NA, NA)),
row.names = c(NA, -20L),
class = "data.frame")
\donttest{
plot_lordif_multiverse(specification_results, items = items_anxiety)
}

}
